# resources/api_doc_template.yaml
swagger: "2.0"
info:
  description: "This API documentation details the resources for the Code Stroke Backend."
  version: "$VERSION"
  title: "Code Stroke Backend"
host: "$HOST"
basePath: "/"
...
schemes:
- "https"
paths:
  /clinicians/register/:
    post:
      tags:
      - "Clinicians"
      summary: "Registers a new clinician."
      description: "Registers a clinician with the backend. This is only generally called by an admin during the initial setup, and would not ordinarily be called by a frontend client. When this request is made, an email is sent to the clinician's email (provided in the request) and you must access the email for a QR code to progress further. The authentication header must contain the username and password of an administrator, not a clinician."
      operationId: "registerClinician"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Basic clinician details."
        required: true
        schema:
          $ref: "#/definitions/CLINICIANS"
      responses:
        200:
          description: "OK"
          schema: 
            type: "object"
            properties:
              success: 
                type: "boolean"
                default: True
              destination:
                type: "string"
        500:
          description: "Error not specified. Please contact the developers."
      security:
        - requiresAdmin: []
  /clinicians/pair/:
    post:
      tags:
      - "Clinicians"
      summary: "Returns a secret the first time it is called."
      description: "When a clinician is registered and receives their QR code, they are required to send the information contained in the QR code to this resource. If it is the first time that clinician calls this resource, then a shared secret will be returned - otherwise, this will return an error."
      operationId: "pairClinician"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "QR code data sent to clinician's email upon registration."
        required: true
        schema: 
          type: "object"
          required: 
            - username
            - password
            - pairing_code
            - backend_domain
            - backend_id
          properties:
            username: 
              type: "string"
            password:
              type: "string"
            pairing_code:
              type: "string"
            backend_domain:
              type: "string"
            backend_id:
              type: "string"
          
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              shared_secret:
                type: "string"
              token:
                type: "string"
        400:
          description: "Request parameters were incomplete."
          schema: 
            type: "object"
            properties:
              success:
                 type: "boolean"
                 default: False
              error_type:
                type: "string"
                enum: ["request"]
        401:
          description: "QR code details did not match server's records."
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                default: False
              error_type:
                type: "string"
                enum: ["checkpoint"]
  /clinicians/set_password/:
    post:
      tags:
      - "Clinicians"
      summary: "Sets the clinician's initial password."
      description: "Sets the clinician's initial password after the pairing step. This will only work the first time it is called."
      operationId: "setPassword"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Initial password details."
        required: true
        schema:
          type: "object"
          required:
            - new_password
          properties:
            new_password:
              type: "string"
      responses:
        200:
          description: "Password set."
          schema:
            type: "object"
            properties:
              success:
                type: boolean
                default: True
        400:
          description: "Body parameters were invalid."
          schema:
            type: "object"
            properties:
              success:
                type: boolean
                default: False
              error_type:
                type: string
                enum: ['request']
      security:
        - requiresClinician: []
  /clinicians/profile/:
    get:
      tags:
        - "Clinicians"
      summary: "Returns basic clinician details."
      description: "Returns some basic profile information of a clinician who accesses this resource with their login credentials."
      operationId: "getProfile"
      produces: 
        - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              signoff_first_name:
                type: string
              signoff_last_name:
                type: string
              signoff_username:
                type: string
              signoff_role:
                type: string
                enum: ['paramedic', 'ed_clinician', 'radiographer', 'stroke_team', 'radiologist', 'stroke_ward', 'neuroint', 'angio_nurse', 'anaesthetist', 'admin']
        401:
          description: "Authentication credentials did not pass."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
      security:
        - requiresClinician: []
...
securityDefinitions:
  requiresClinician:
    type: "basic"
    description: "This is the standard 2FA method for a clinician. The time-based token based on the shared secret should be concatenated onto the password with a colon i.e. `password:token` - if using the headers directly, the whole string should be of the form `username:password:token` before being base64 encoded."
  requiresAdmin:
    type: "basic"
    description: "This route requires administrator credentials, which are just a standard username and password. These resources should not generally be called from the ordinary frontend clients and may be deprecated in the future."
definitions:
  FailedAuthClinician:
    type: "object"
    properties:
      success: 
        type: boolean
        default: False
      error_type:
        type: string
        enum: ["auth"]